# Conflicts:
#	src/mc/certificates/identity/PlayerAuthenticationInfo.h
#	src/mc/client/gui/screens/ScreenEventData.h
#	src/mc/common/Globals.h
#	src/mc/deps/core/file/Path.h
#	src/mc/deps/core/file/PathPart.h
#	src/mc/deps/core/file/PathView.h
#	src/mc/deps/core/math/Color.h
#	src/mc/deps/core/math/Vec3.h
#	src/mc/deps/core/string/HashedString.h
#	src/mc/deps/core_graphics/ImageDescription.h
#	src/mc/deps/json/Value.h
#	src/mc/deps/vanilla_components/HitboxComponent.h
#	src/mc/entity/components_json_legacy/CombatRegenerationComponent.h
#	src/mc/nbt/CompoundTag.h
#	src/mc/nbt/CompoundTagVariant.h
#	src/mc/nbt/ListTag.h
#	src/mc/network/NetworkIdentifier.h
#	src/mc/network/packet/ModalFormRequestPacketPayload.h
#	src/mc/network/packet/ServerSettingsResponsePacketPayload.h
#	src/mc/network/packet/StopSoundPacketPayload.h
#	src/mc/network/packet/TextPacket.h
#	src/mc/network/packet/TransferPacketPayload.h
#	src/mc/network/packet/serialize/SerializedPayloadPacket.h
#	src/mc/platform/CallStack.h
#	src/mc/platform/threading/PrioritizeSharedOwnership.h
#	src/mc/safety/RedactableString.h
#	src/mc/server/commands/CommandOutput.h
#	src/mc/server/commands/CommandOutputMessage.h
#	src/mc/util/MolangVariableMap.h
#	src/mc/world/actor/Actor.h
#	src/mc/world/actor/SynchedActorDataEntityWrapper.h
#	src/mc/world/item/ItemInstance.h
#	src/mc/world/item/ItemStack.h
#	src/mc/world/level/BlockPos.h
#	src/mc/world/level/dimension/DerivedDimensionArguments.h

+++ src-server/mc/client/gui/LiteralType.h
@@ -2,4 +2,4 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class LiteralType : int {};
+enum class LiteralType : uchar {};
